rand: introduce a pseudo random number generator for bare targets
This is a PRNG meant to be used in a Proxy-WASM filter (but it can
be extracted to receive a timing function) so that we don't need
to go the WASI route to have platform support for RNGs.

With this we'll have a thread-local PRNG that is seeded from the
Proxy-WASM current time function (which in Envoy is currently not
precise enough, so entropy is reduced but still useful) by taking
its jitter (rand_jitter crate) and hashing it (rand_seeder crate)
before using the resulting value as the seed for any one of an
initial choice of pseudo RNGs (PCG32, Xoshiro128**, XorShift).

The implementation relies on a thread-local RefCell which holds
the actual PRNG data, which needs to be initialized at some point.

Initializing the PRNG is not enforced at compile time because there
are limitations in the way we can assign TLS values that would end
up translating the ergonomic issues down the interface (ie. you'd
be unwrapping/checking constantly for something that ideally happens
only once per thread, which is to say only once ever, since this
WASM environment is single-threaded).

The next commit shall include the initialization for our filter.