root_context: initialize and test the PRNG
The PRNG does not need (at all) to be part of the context, but we do
this here because we are going to make use of it soon for identifying
specific threads. That is, you can use the PRNG from other contexts,
and even outside any context at all, as long as you ensure the PRNG is
initialized somewhere before you try using it.