operations: refactor and add stack and string operations
This introduces a series of useful operations to provide for
some flexibility in looking up values and operating on them
to find credentials.

The concept of a stack of values is introduced, with operations
working on single values which will pop the last value and
generate a new one, and operations working on the whole stack,
to reverse values, join them, selecting a few, or ensuring
a certain number of values are present.

String operations are also introduced so you can now reverse
strings, replace substrings, ensure they have a certain length,
and other operations.

In the future we might want to apply some uniformity to these
operations so that we just receive a given input (ie. headers,
query string parameters) and we would apply a pipeline of
lookups and transformations to obtain the values without an
explicit split in between the initial value lookups and the
rest of the operations.