operation/control: add conditional control primitives
This allows just enough control operations to set up conditionals
without introducing loops to the control flow.

These control flow operations always pass the original input to any
contained operations, but differ on the logic on when the operation
is considered successful as well as on whether the operations are
mutable or not.

`True` is an always successful operation, whereas `False` is always
unsuccessful.

`Any` is essentially a short-circuit `or` operation, whereas our
`OneOf` is as short-circuit `xor`. Both operations, when successful,
return the output of the single successful operation.

The rest of the operations just evaluate the operations and drop any
outputs, returning the input when succesful.

`All` requires that all operations are successful, and if so it returns
the input. Conversely, `None` requires that all operations fail, and if
so it returns the input.

`Assert` just ensures that the operation is successful and passes on its
input in that case. Conversely, `Refute` passes on the input if the
operation fails.

`Log` does not change control, but does not seem to merit yet its own
category. It just outputs a given message.